[librpbase] IAesCipher: Various cleanups.
- Changed the four-argument decrypt() function to be an inline wrapper
  around the two-argument decrypt() function in IAesCipher.hpp. I was
  originally going to remove it, but it's used by WiiPartition, so
  it's staying as a wrapper.
- Use `size_t` instead of `unsigned int` for consistency with other
  functions that take pointers and sizes.
- Rename the parameters to `pKey`, `pIV`, `size`, etc.
- AesCAPI_NG::setKey(): Removed the incorrect `d->key != pKey`
  comparison, since this is a pointer comparison and it will
  always be true.

+++ src/librpbase/crypto/KeyManager.cpp
@@ -405,7 +405,7 @@ KeyManager::VerifyResult KeyManager::getAndVerify(const char *keyName, KeyData_t
 	// make a temporary copy.
 	unique_ptr<uint8_t[]> tmpData(new uint8_t[verifyLen]);
 	memcpy(tmpData.get(), pVerifyData, verifyLen);
-	unsigned int size = cipher->decrypt(tmpData.get(), verifyLen);
+	size_t size = cipher->decrypt(tmpData.get(), verifyLen);
 	if (size != verifyLen) {
 		// Decryption failed.
 		return VERIFY_IAESCIPHER_DECRYPT_ERR;
